Extracted branding, extracted questions into a questions wrapper, created file and folder for admin and user views, but have not fleshed them out.
