+++ src/components/branding/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import ArrowBackIosSharpIcon from '@material-ui/icons/ArrowBackIosSharp';
+
+export interface BrandingProps {}
+
+const Branding: React.FC = () => {
+  return <h1>{`De>_Support`}</h1>;
+};
+
+export default Branding;
