+++ src/view/Root/index.tsx
@@ -1,51 +1,16 @@
 import React from 'react';
-import {
-  Button,
-  makeStyles,
-  Paper,
-  TextField,
-  Accordion,
-} from '@material-ui/core';
-//TESTING DATA
-import questions from '../../data';
-//TESTING DATA
-import QuestionComponent from '../../components/questions_component';
+import Branding from '../../components/branding';
+import QuestionsWrapper from '../../components/questions_wrapper';
 import QuestionSubmission from '../../components/questions-submission';
-import ArrowBackIosSharpIcon from '@material-ui/icons/ArrowBackIosSharp';
 
 interface ViewProps {}
 
-const useStyles = makeStyles((theme) => ({
-  rootPaper: {
-    display: 'flex',
-    flexDirection: 'column',
-    flexWrap: 'wrap',
-    '& > div': {
-      marginBottom: theme.spacing(2),
-      padding: theme.spacing(1),
-      minHeight: theme.spacing(16),
-    },
-    chip: {
-      margin: theme.spacing(0.5),
-    },
-  },
-}));
-
 const RootView: React.FC = (props: ViewProps) => {
-  const classes = useStyles();
   return (
     <>
-      <h1>
-        {`De`} <ArrowBackIosSharpIcon fontSize="inherit" color="error" />
-        {`_Support`}
-      </h1>
+      <Branding />
       <QuestionSubmission />
-      <div className={classes.rootPaper}>
-        <h2>Question List</h2>
-        {questions.map((q, i) => {
-          return <QuestionComponent q={q} i={i} />;
-        })}
-      </div>
+      <QuestionsWrapper />
     </>
   );
 };
